pip-requirements.txt
django>=1.3
easy-thumbnails
django-model-utils

# for tests
-e git+https://github.com/nose-devs/nose.git#egg=nose
django-nose
coverage